- feature (detailed-video.screen): Added the connectivity check service to this screen. - feature (connectivity-check.service): Added the logic of the connectivity service.
